FIX: Fixed that the tag deleter didnt delete non hierarchical tags

+++ src/ui/obsidian-ui-components/item-views/review-queue-list-view.tsx
@@ -340,26 +340,29 @@ export class ReviewQueueListView extends ItemView {
 
             fileMenu.addSeparator();
 
-            fileMenu.addItem((item) => {
-                item.setTitle(t("DELETE_NOTE_SCHEDULING_DATA_IN_NOTE"))
-                    .setIcon("trash")
-                    .setWarning(true)
-                    .onClick(async () => {
-                        new ConfirmationModal(
-                            this.plugin.app,
-                            t("DELETE_NOTE_SCHEDULING_DATA_IN_NOTE"),
-                            t("CONFIRM_NOTE_SCHEDULING_DATA_IN_NOTE_DELETION"),
-                            t("NOTE_SCHEDULING_DATA_IN_NOTE_DELETION_IN_PROGRESS"),
-                            () => {
-                                deleteNoteSchedulingDataInNote(
-                                    file,
-                                    this.plugin.data.settings.deleteTagsOnSchedulingDataDeletion,
-                                    this.plugin.data.settings.tagsToReview,
-                                );
-                            },
-                        ).open();
-                    });
-            });
+            if (this.plugin.data.settings.showDeleteButtonInFileMenu) {
+                fileMenu.addItem((item) => {
+                    item.setTitle(t("DELETE_NOTE_SCHEDULING_DATA_IN_NOTE"))
+                        .setIcon("trash")
+                        .setWarning(true)
+                        .onClick(async () => {
+                            new ConfirmationModal(
+                                this.plugin.app,
+                                t("DELETE_NOTE_SCHEDULING_DATA_IN_NOTE"),
+                                t("CONFIRM_NOTE_SCHEDULING_DATA_IN_NOTE_DELETION"),
+                                t("NOTE_SCHEDULING_DATA_IN_NOTE_DELETION_IN_PROGRESS"),
+                                () => {
+                                    deleteNoteSchedulingDataInNote(
+                                        file,
+                                        this.plugin.data.settings
+                                            .deleteTagsOnSchedulingDataDeletion,
+                                        this.plugin.data.settings.tagsToReview,
+                                    );
+                                },
+                            ).open();
+                        });
+                });
+            }
 
             fileMenu.showAtPosition({
                 x: event.pageX,
